backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
sqlite
chromadb
ollama
sentence-transformers
python-multipart
python-dotenv
aiofiles
markdown
beautifulsoup4
pytest
httpx
